requirements.txt
fastapi
uvicorn
requests
PyMuPDF
groq
python-dotenv
python-multipart
openai
pdfplumber
pydantic